requirements.txt
fastapi
uvicorn
pyyaml

transformers
torch
accelerate

protobuf
sentencepiece